backend/requirements.txt
playwright
jinja2
matplotlib
fastapi
uvicorn
sqlalchemy
selectolax
apscheduler
pydantic
python-dotenv
curl_cffi
reportlab
svglib
psycopg2-binary
groq
twilio
httpx
